arrumando logica do header, adicionando alertas de criacao de treino/login/logout

+++ befit-app/src/components/Header/Header.jsx
@@ -5,6 +5,8 @@ import React, { useState } from "react";
 import { BrowserRouter as Router, Routes, Route, useNavigate, Link, NavLink } from 'react-router-dom'
 import axios from 'axios';
 import { Component } from 'react';
+import swal from 'sweetalert';
+
 function Header(props) {
 
 
@@ -17,7 +19,6 @@ function Header(props) {
     switch (pageAtual) {
       case '/usuario/exercicios':
         return false
-
 
       case '/usuario/meustreinos':
         return false
@@ -33,6 +34,7 @@ function Header(props) {
       case '/usuario/minhasdietas':
         return false
 
+
       default:
         return true
 
@@ -40,45 +42,49 @@ function Header(props) {
 
   }
 
-  function headerSair(){
+  function headerSair() {
 
-    switch(pageAtual){
-
-    case '/usuario/exercicios':
-      return true
-
-    case '/usuario/meustreinos':
-      return true  
-
-    case '/usuario/dietas':
-      return true  
-    
-    case '/usuario/minhasdietas':
-      return true    
-
-    default:
-      return false
-    }
-  }
-
-  function headerVoltar(){
     switch (pageAtual) {
-      case'/usuario/criar/treinos':
+
+      case '/usuario/exercicios':
         return true
 
-      case '/treino/*':
-        return true 
-    
+      case '/usuario/meustreinos':
+        return true
+
+      case '/usuario/dietas':
+        return true
+
+      case '/usuario/minhasdietas':
+        return true
+
       default:
         return false
     }
   }
 
-  function headerPagHome(){
+  function headerVoltar() {
+    switch (pageAtual) {
+      case '/usuario/criar/treinos':
+        return true
+
+      case `/treino/${sessionStorage.getItem("idTreino")}`:
+        console.log(`/treino/${sessionStorage.getItem("idTreino")}`)
+        return true
+
+      default:
+        return false
+    }
+  }
+
+  function headerPagHome() {
     switch (pageAtual) {
       case '/usuario/exercicios':
         return false
 
+      case `/treino/${sessionStorage.getItem("idTreino")}`:
+        console.log(`/treino/${sessionStorage.getItem("idTreino")}`)
+        return false
 
       case '/usuario/meustreinos':
         return false
@@ -91,7 +97,7 @@ function Header(props) {
       case '/usuario/dietas':
         return false
 
- case '/usuario/minhasdietas':
+      case '/usuario/minhasdietas':
         return false
 
       case !'/treino/*':
@@ -120,10 +126,11 @@ function Header(props) {
       case '/usuario/dietas':
         return true
 
- case '/usuario/minhasdietas':
+      case '/usuario/minhasdietas':
         return true
 
-      case '/treino/*':
+      case `/treino/${sessionStorage.getItem("idTreino")}`:
+        console.log(`/treino/${sessionStorage.getItem("idTreino")}`)
         return true
 
       default:
@@ -138,13 +145,13 @@ function Header(props) {
         return true
 
       case '/usuario/meustreinos':
-        return true  
+        return true
 
       case '/usuario/dietas':
-        return true  
-      
+        return true
+
       case '/usuario/minhasdietas':
-        return true    
+        return true
 
       default:
         return false
@@ -156,22 +163,39 @@ function Header(props) {
 
 
   const handleLogOut = () => {
-    axios.patch(`http://localhost:8080/usuarios/logout/${sessionStorage.getItem("personId")}`)
-      .then((res) => {
-        console.log(res)
-        var autenticado = sessionStorage.setItem("autenticado", false)
 
-        console.log('usuario autenticado? ', `${sessionStorage.getItem("autenticado")}`)
-        if (!autenticado) {
-          sessionStorage.clear()
-          navigate('/')
-          alert('usuario deslogado')
+    swal({
+      title: "Tem certeza que gostaria de sair?",
+      text: "Você será deslogado do nosso site ",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          axios.patch(`http://localhost:8080/usuarios/logout/${sessionStorage.getItem("personId")}`)
+            .then((res) => {
+              console.log(res)
+              var autenticado = sessionStorage.setItem("autenticado", false)
+              console.log('usuario autenticado? ', `${sessionStorage.getItem("autenticado")}`)
+              if (!autenticado) {
+                sessionStorage.clear()
+                navigate('/')
+              } 
+            })
+            .catch(err => {
+              console.log(err)
+              console.log('erro no .logout')
+            })
+          swal("Poof! Você foi deslogado!", {
+            icon: "success",
+          });
+        } else {
+          swal("Obrigado por continuar no nosso site!");
         }
-      })
-      .catch(err => {
-        console.log(err)
-        console.log('erro no .logout')
-      })
+      });
+
+
   }
 
   return (<>
@@ -185,7 +209,7 @@ function Header(props) {
               {headerChange() && headerPagHome() && <li id='home'> <NavLink to="/" end> Home</NavLink></li>}
               {/* <li> <NavLink to="/exerciciosHome" activeClassName="active"> <a>Exercícios</a></NavLink></li> */}
               {headerPagHome() && headerChange() && <li id='sobre'> <NavLink to="/sobre" >Sobre</NavLink></li>}
-              {headerPagHome() && <li id='comecar'> <NavLink to="/signin"> Começar </NavLink></li>}   
+              {headerPagHome() && <li id='comecar'> <NavLink to="/signin"> Começar </NavLink></li>}
               {headerChangeUserPage() && headerVoltar() && <li> <button className='btn-voltar-pag' onClick={() => navigate(-1)}>Voltar</button></li>}
               {headerChangeUserPage() && headerHotSite() && <li>  <span className='user'>Olá {sessionStorage.getItem("nome")}!</span></li>}
               {headerHotSite() && <li> <button className='btn-hot-site'>  <NavLink to="/hotsite">HotSite</NavLink></button></li>}
